revert multistage strategy
